Restructuring Mr Mustard states.py, transformations.py and circuit_components_utils.py + pylint misc (#444)
