+++ mrmustard/math/caching.py
@@ -15,8 +15,10 @@
 """This module contains the logic for cachin tensor functions in Mr Mustard."""
 
 from functools import lru_cache, wraps
-from mrmustard.math.backend_manager import BackendManager
+
 import numpy as np
+
+from mrmustard.math.backend_manager import BackendManager
 
 math = BackendManager()
