+++ tests/test_lab_dev/test_states/test_vacuum.py
@@ -1,0 +1,45 @@
+# Copyright 2023 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Tests for the ``Vacuum`` class."""
+
+# pylint: disable=protected-access, unspecified-encoding, missing-function-docstring, expression-not-assigned, pointless-statement
+
+import numpy as np
+import pytest
+
+from mrmustard import math
+from mrmustard.lab_dev.states import Vacuum
+
+
+class TestVacuum:
+    r"""
+    Tests for the ``Vacuum`` class.
+    """
+
+    @pytest.mark.parametrize("modes", [(0,), (0, 1), (3, 19, 2)])
+    def test_init(self, modes):
+        state = Vacuum(modes)
+
+        assert state.name == "Vac"
+        assert list(state.modes) == sorted(modes)
+        assert state.n_modes == len(modes)
+
+    @pytest.mark.parametrize("n_modes", [1, 3])
+    def test_representation(self, n_modes):
+        rep = Vacuum(range(n_modes)).representation
+
+        assert math.allclose(rep.A, np.zeros((1, n_modes, n_modes)))
+        assert math.allclose(rep.b, np.zeros((1, n_modes)))
+        assert math.allclose(rep.c, [1.0])
